Added some simple visuals for debugging the trackball(commented). Solved some minor bugs. Found a bug with the trackball and camera coordinate systems.

+++ include/inviwo/core/interaction/trackball.h
@@ -20,7 +20,7 @@ namespace inviwo {
     private:
         CameraProperty* camera_;
 
-        ivec2 lastMousePos_;
+        vec2 lastMousePos_;
         vec3 lastTrackballPos_;
 
         vec3 mapNormalizedMousePosToTrackball(vec2 mousePos);
